requirements.txt
fastapi
pydantic
uvicorn
pyautogen
python-dotenv
requests
openai
azure-identity